% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_code_generator.R
\name{pin_code_generator}
\alias{pin_code_generator}
\title{Code generator for pin calculation}
\usage{
pin_code_generator(
  cluster_questions_df,
  cluster_name,
  col_name_indicator,
  col_name_severity_1,
  col_name_severity_2,
  col_name_severity_3,
  col_name_severity_4,
  col_name_severity_5
)
}
\arguments{
\item{cluster_questions_df}{dataframe input}

\item{cluster_name}{Cluster name}

\item{col_name_indicator}{Column name of the indicator description}

\item{col_name_severity_1}{Column name of the level 1 sevirity score}

\item{col_name_severity_2}{Column name of the level 2 sevirity score}

\item{col_name_severity_3}{Column name of the level 3 sevirity score}

\item{col_name_severity_4}{Column name of the level 4 sevirity score}

\item{col_name_severity_5}{Column name of the level 5 sevirity score}
}
\value{
Return a ready to use sckeleton code. The code is copied into the clipboard. Use ctrl+V / cmd+V to paste it
}
\description{
Code generator for pin calculation
}
\examples{
print("Code example placeholder")
}
